GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - main (refs/gitbutler/main)
   branch head: c38ebd6195f4e52f406743da07ca523956bf03c2
   - .gitignore
   - Tempora.nnwtheme/Info.plist
   - Tempora.nnwtheme/template.html
   - Tempora.nnwtheme/stylesheet.css

Your previous branch was: refs/heads/main

The sha for that commit was: d527614616b8386368eac401228ec6b54858d393

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch